if agent eats food fix

+++ Direction.py
@@ -12,7 +12,7 @@ class Direction(Enum):
 # Point = namedtuple('Point', 'x, y')
 class Point(object):
     __slots__ = ('x', 'y')
-    def __init__(self, x: int, y: int):
+    def __init__(self, x: float, y: float):
         self.x = x
         self.y = y
